+++ CesiumGeometry/include/CesiumGeometry/IntersectionTests.h
@@ -29,7 +29,8 @@ public:
   rayPlane(const Ray& ray, const Plane& plane) noexcept;
 
   /**
-   * @brief Determines whether the point is completely inside the triangle.
+   * @brief Determines whether a given point is completely inside a triangle
+   * defined by three 2D points.
    *
    * @param point The point to check.
    * @param triangleVertA The first vertex of the triangle.
@@ -37,14 +38,15 @@ public:
    * @param triangleVertC The third vertex of the triangle.
    * @return Whether the point is within the triangle.
    */
-  static bool pointInTriangle2D(
+  static bool pointInTriangle(
       const glm::dvec2& point,
       const glm::dvec2& triangleVertA,
       const glm::dvec2& triangleVertB,
       const glm::dvec2& triangleVertC) noexcept;
 
   /**
-   * @brief Determines whether the point is completely inside the triangle.
+   * @brief Determines whether a given point is completely inside a triangle
+   * defined by three 3D points.
    *
    * @param point The point to check.
    * @param triangleVertA The first vertex of the triangle.
@@ -59,8 +61,9 @@ public:
       const glm::dvec3& triangleVertC) noexcept;
 
   /**
-   * @brief Determines whether the point is completely inside the triangle and
-   * outputs the barycentric coordinates for the point.
+   * @brief Determines whether the point is completely inside a triangle defined
+   * by three 3D points. If the point is inside, this also outputs the
+   * barycentric coordinates for the point.
    *
    * @param point The point to check.
    * @param triangleVertA The first vertex of the triangle.